Move Tracker from trackjdwp.ts to separate file
